fix: big decimal to double

+++ app/src/main/java/fr/henri/potech/bookshop/domain/Book.kt
@@ -1,6 +1,5 @@
 package fr.henri.potech.bookshop.domain
 
-import java.math.BigDecimal
 import java.net.URL
 
 data class Book(
@@ -8,7 +7,7 @@ data class Book(
     val title: String,
     val isbn: String,
     val synopsis: String,
-    val price: BigDecimal
+    val price: Double
 ) {
     companion object {}
 }
